Custom clipping
Multiple prompts in the console now. Can select if the whole video should be clipped with a designated length of each clip. Can also say if a clip should be grabbed from a whole video with the start and end time of the clip. Also added custom naming of the clips for organization purposes